fix(vfs): fastfetch was showing wrong disk size and it was static docs(security): removed af_unix

+++ kernel/src/fs/vfs.h
@@ -68,20 +68,22 @@ typedef uint64_t (*mmap_type_t)(struct vfs_node *, uint64_t addr,
 typedef int (*poll_type_t)(struct vfs_node *, int events);
 typedef int (*fallocate_type_t)(struct vfs_node *, int mode, uint32_t offset,
                                 uint32_t len);
-// statfs structure (Linux x86_64 compatible)
+// statfs structure (Linux x86_64 ABI compatible)
+// All fields are 'long' (8 bytes on x86_64).
+// f_fsid is __kernel_fsid_t = int[2] = exactly 8 bytes — NOT uint64_t[2].
 struct statfs_buf {
-  uint64_t f_type;
-  uint64_t f_bsize;
-  uint64_t f_blocks;
-  uint64_t f_bfree;
-  uint64_t f_bavail;
-  uint64_t f_files;
-  uint64_t f_ffree;
-  uint64_t f_fsid[2];
-  uint64_t f_namelen;
-  uint64_t f_frsize;
-  uint64_t f_flags;
-  uint64_t f_spare[4];
+  int64_t  f_type;
+  int64_t  f_bsize;
+  int64_t  f_blocks;
+  int64_t  f_bfree;
+  int64_t  f_bavail;
+  int64_t  f_files;
+  int64_t  f_ffree;
+  int32_t  f_fsid[2];   /* __kernel_fsid_t: two int32 = 8 bytes total */
+  int64_t  f_namelen;
+  int64_t  f_frsize;
+  int64_t  f_flags;
+  int64_t  f_spare[4];
 };
 
 typedef int (*statfs_type_t)(struct vfs_node *, struct statfs_buf *buf);
